refactor entities int to int64

+++ backend/internal/domain/entities/deck.go
@@ -2,6 +2,6 @@ package entities
 
 type Deck struct {
 	Id         int
-	SessionId  int
+	SessionId  int64
 	Flashcards []FlashCard
 }
